modules/uda-location-producer/requirements.txt
kafka-python==2.0.2
SQLAlchemy==1.3.19
grpcio==1.40.0
grpcio-tools==1.40.0
protobuf==3.17.3